Add unit test on json conversion, address peer review
